requirements.txt
opencv-python-headless
tensorflow==2.8.0
pandas==1.2.3
keras==2.8.0
Keras-Preprocessing==1.1.2
numpy==1.21.5
Pillow==9.0.1
click<=8.0.4
streamlit
scipy==1.7.3
matplotlib==3.5.1
